Moved caffe uploading part to a separate file.
