+++ webpack-stats.dev.json
@@ -1,1 +1,1 @@
-{"status":"done","publicPath":"http://localhost:3000/","chunks":{"main":[{"name":"static/js/bundle.js","publicPath":"http://localhost:3000/static/js/bundle.js","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\static\\js\\bundle.js"},{"name":"0.7dbfc768b0a6f5697945.hot-update.js","publicPath":"http://localhost:3000/0.7dbfc768b0a6f5697945.hot-update.js","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\0.7dbfc768b0a6f5697945.hot-update.js"},{"name":"static/js/bundle.js.map","publicPath":"http://localhost:3000/static/js/bundle.js.map","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\static\\js\\bundle.js.map"},{"name":"0.7dbfc768b0a6f5697945.hot-update.js.map","publicPath":"http://localhost:3000/0.7dbfc768b0a6f5697945.hot-update.js.map","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\0.7dbfc768b0a6f5697945.hot-update.js.map"}]},"error":"ModuleError","message":"\n  Line 98:  Parsing error: Unterminated JSX contents\n\n   96 | \n   97 |       </div>\n>  98 |       </div>\n      |             ^\n   99 |     );\n  100 |   }\n  101 | }\n\n"}
+{"status":"done","publicPath":"http://localhost:3000/","chunks":{"main":[{"name":"static/js/bundle.js","publicPath":"http://localhost:3000/static/js/bundle.js","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\static\\js\\bundle.js"},{"name":"0.a984713817ed6174bbac.hot-update.js","publicPath":"http://localhost:3000/0.a984713817ed6174bbac.hot-update.js","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\0.a984713817ed6174bbac.hot-update.js"},{"name":"static/js/bundle.js.map","publicPath":"http://localhost:3000/static/js/bundle.js.map","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\static\\js\\bundle.js.map"},{"name":"0.a984713817ed6174bbac.hot-update.js.map","publicPath":"http://localhost:3000/0.a984713817ed6174bbac.hot-update.js.map","path":"C:\\Users\\Ollie\\Projects\\Uni\\deep_introspection\\frontend\\0.a984713817ed6174bbac.hot-update.js.map"}]},"error":"ModuleError","message":"\n  Line 45:  Parsing error: Unexpected token, expected }\n\n  43 |             >\n  44 |              {(this.props.first) ? <p style={{paddingLeft:15, paddingRight: 15, paddingTop: 15}}>Before using Deep Introspection, you need to upload a network model to analyse.</p> : <div/>}\n> 45 |           <UploadCaffe onRef={ref => (this.child = ref) updateProgress={(percentage) => this.setState({percentage:percentage})} callbackParent={(model) => this.props.callbackParent(model)}/>\n     |                                                         ^\n  46 |         <div className=\"pt-dialog-footer\" >\n  47 |                     {(this.state.percentage != null) ?\n  48 |                       <div class=\"pt-progress-bar pt-intent-primary\" style={{width: \"65%\"}}>\n\n"}
